+++ MovieStore/MovieStore/Movie.cpp
@@ -2,10 +2,24 @@
 
 Movie::Movie()
 {
+	Item('m', 0, "");
+	director = "";
+	releaseYear = 0;
+	type = "";
+}
+
+Movie::Movie(char id, int stock, const string &dir, const string &name,
+	int year, const string & type)
+{
+	Item(id, stock, name);
+	director = dir;
+	releaseYear = year;
+	this->type = type;
 }
 
 Movie::~Movie()
 {
+
 }
 
 string Movie::getDirector() const
@@ -27,3 +41,4 @@ string Movie::getTitle() const
 {
 	return title;
 }
+
